CurrencyApi voltou a estar no index
-- Agora que a currencyApi é um módulo, não há mais necessidade de existir um index para instanciá-la e um arquivo para definí-la, então o código do módulo foi motivo para seu índice
-- Adaptado o sistema para essa mudança
