+++ src/db/models/transaction.ts
@@ -1,7 +1,7 @@
 import mongoose, { Schema, Document } from 'mongoose'
 import { ObjectId, Decimal128 } from 'mongodb'
 import { Person } from './person/interface'
-import { SuportedCurrencies } from '../../currencyApi/currencyApi'
+import type { SuportedCurrencies } from '../../currencyApi'
 
 /** Interface base de uma transaction */
 interface Transaction {
